requirements.txt
pyirsdk
PyQt5 